Add End2End emulated Android testing
This adds an end2end test which runs on an emulated (or physical)
android device and exercises the full app and networking functionality,
simulating actual usage and asserting correct network conditions.

It updates the app to include 2 additional simple input fields which
allow for the user (and automated test) to set the proxy IP and port
fields.

It also includes a docker-compose.yml for starting a containerized h2o
proxy as well as a simple echo server which we use for the test to
assert that upon enabling the relay service we've changed IPs.